Port to Acorn 6.0
