+++ test/test.js
@@ -1,11 +1,12 @@
 "use strict"
 
 const assert = require("assert")
-const acorn = require("..")
+const acorn = require("acorn")
+const Parser = acorn.Parser.extend(require(".."))
 
 function test(text, expectedResult, additionalOptions) {
   it(text, function () {
-    const result = acorn.parse(text, Object.assign({ ecmaVersion: 9, plugins: { bigInt: true } }, additionalOptions))
+    const result = Parser.parse(text, Object.assign({ ecmaVersion: 9 }, additionalOptions))
     assert.deepEqual(result.body[0], expectedResult)
   })
 }
@@ -13,7 +14,7 @@ function testFail(text, expectedError, additionalOptions) {
   it(text, function () {
     let failed = false
     try {
-      acorn.parse(text, Object.assign({ ecmaVersion: 9, plugins: { bigInt: true } }, additionalOptions))
+      Parser.parse(text, Object.assign({ ecmaVersion: 9 }, additionalOptions))
     } catch (e) {
       assert.equal(e.message, expectedError)
       failed = true
